Pull bar chart data formatting into functions in utils file
